% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{alignment}
\alias{alignment}
\title{Alignment}
\usage{
alignment(spectra, baseline.threshold = 0.02, reference = NULL,
  max.shift = 0.002)
}
\arguments{
\item{spectra}{Data frame with spectra in columns and chemical shift in rows.
Colnames of this data frame correspond to pure metabolite names and rownames
to chemical shift grid (in p.p.m).}

\item{baseline.threshold}{Value of baseline threshold used to identify peaks.
Default to 0.02.}

\item{reference}{Index of reference spectra. Default to \code{NULL},
\emph{i.e.} reference spectrum is detect automatically with the
\code{findRef} function of package \code{speaq}}

\item{max.shift}{Maximum shift allowed for the alignment. Default to 0.002.}
}
\value{
A data frame with aligned spectra in columns and chemical shifts
(in p.p.m.) in rows.
}
\description{
Align spectra of a data frame with the CluPA algorithm (Vu et al., (2011))
}
\examples{
current_path <- file.path(system.file("extdata", package = "ASICS"),
                          "spectra_example.txt")
spectra_data <- read.table(current_path, header = TRUE, row.names = 1)
spectra_base_cor <- alignment(spectra_data)
}
\references{
Vu, T. N., Valkenborg, D., Smets, K., Verwaest, K. A., Dommisse,
R., Lemiere, F., ... & Laukens, K. (2011). An integrated workflow for robust
alignment and simplified quantitative analysis of NMR spectrometry data.
\emph{BMC bioinformatics}, \strong{12}(1), 405.
}
