% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS.R, R/ASICS_new.R
\name{ASICS}
\alias{ASICS}
\alias{ASICS}
\title{Automatic Statistical Identification in Complex Spectra}
\usage{
ASICS(path, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2, nrow = 1),
  max.shift = 0.02, which.spectra = "last", library.metabolites = NULL,
  threshold.noise = 0.02)

ASICS(path, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2, nrow = 1),
  max.shift = 0.02, which.spectra = "last", library.metabolites = NULL,
  threshold.noise = 0.02)
}
\arguments{
\item{path}{folder path of the Bruker files}

\item{exclusion.areas}{exclusion areas of the quantification}

\item{max.shift}{maximum chemical shift allowed (in ppm)}

\item{which.spectra}{if more than one spectra by sample, spectra to choose
(either "first", "last" or its number)}

\item{library.metabolites}{path of the library of standard if not the default
one}

\item{threshold.noise}{threshold noise}

\item{name}{folder path of the Bruker files}

\item{ZoneAEnlever}{exclusion areas to remove before the quantification}

\item{DecalageMax}{maximum chemical shift allowed}

\item{DecalageLib}{chemical shift between library of pure spectra and complex mixture}

\item{sample}{if more than one spectra by sample, spectra to choose (either "first",
"last" or its number}

\item{seed}{seed}

\item{libraryMetab}{path of the library of standard if not the default one}

\item{seuilBruit}{threshold noise}
}
\description{
Description

Description
}
\examples{
result <- ASICS(name = "./spectres_exemple/AG_faq_Beck01",
 ZoneAEnlever = matrix(c(4.5,5.1,5.5,6.5), ncol = 2, byrow = TRUE),
 DecalageMax = 0.02, DecalageLib = 0, sample = "last", seed = 12345,
 libraryMetab = NULL)
result <- ASICS(path = "./spectres_exemple/AG_faq_Beck01",
 exclusion.areas = matrix(c(4.5,5.1,5.5,6.5), ncol = 2, byrow = TRUE),
 max.shift = 0.02, which.spectra = "last", library.metabolites = NULL)
}
\keyword{NMR}
\keyword{metabolites}
\keyword{quantification}
