% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICSResults-class.R
\docType{class}
\name{ASICSResults-class}
\alias{ASICSResults-class}
\title{Class \linkS4class{ASICSResults}}
\description{
Objects of class \linkS4class{ASICSResults} represent results of ASICS for a
set of spectra.
It an extension of class \linkS4class{Spectra} with additional slots for
quantification results, recomposed spectra and deformed library.
}
\section{Slots}{

\describe{
\item{\code{sample.name}}{Character vector of sample names.}

\item{\code{ppm.grid}}{Numeric vector of a unique grid (definition domain) for all
spectra (in ppm).}

\item{\code{spectra}}{Numeric matrix of original spectra in column.
Columns are in the same order as for \code{sample.name} and rows correspond
to points of \code{ppm.grid}.}

\item{\code{recomposed.spectra}}{Numeric matrix of recomposed spectra (in column)
with estimated concentrations. Columns are in the same order as for
\code{sample.name} and rows correspond to points of \code{ppm.grid}.}

\item{\code{quantification}}{Data-frame with identified metabolites and their
relative concentrations.}

\item{\code{deformed.library}}{a list of \linkS4class{PureLibrary} containing the
deformed libraries of each sample.}
}}

\section{Methods}{

  Multiple methods can be applied on \linkS4class{Spectra} objects.
  \itemize{
    \item As usual for S4 object, show and summary methods are available, see
    \link[=summary-methods]{Object summary}
    \item All slots have an accessor \code{get_slot name}, see
    \link[=accessors-methods]{Accessors}
    \item Two objects can be combine or a subset can be extract, see
    \link[=combineAndSubset-methods]{Combine and subset methods}
    \item All spectra contained in an object can be represent in a plot, see
    \link[=visualization-methods-spectra]{Visualization methods}
  }
}

\seealso{
\linkS4class{Spectra}
}
