% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_guide.R
\name{ASICSUsersGuide}
\alias{ASICSUsersGuide}
\title{View ASICS User's Guide}
\usage{
ASICSUsersGuide(view = TRUE)
}
\arguments{
\item{view}{if \code{TRUE}, the user's guide will be opened with default
browser}
}
\value{
Character string giving the file location. If \code{view = TRUE},
the HTML viewer is started and the User’s Guide is opened, as a side effect.
}
\description{
Open the ASICS User's Guide (with default browser)
}
\details{
The function \code{vignette("ASICS")} will find the short
ASICS vignette that describes the main functions and how to obtain the ASICS
User's Guide.
The User's Guide is not itself a true vignette because it is not
automatically generated during the package build process.
If the operating system is not Windows, then the HTML viewer used is the one
given by \code{Sys.getenv("R_BROWSER")}. The HTML viewer can be changed using
\code{Sys.setenv(R_BROWSER = )}.
}
\examples{
# To get the location
ASICSUsersGuide(view = FALSE)

# To open in a HTML viewer
\dontrun{ASICSUsersGuide()}
}
