% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{binning}
\alias{binning}
\title{Binning/Bucketing of NMR spectra}
\usage{
binning(spectra, bin = 0.01, exclusion.areas = matrix(c(4.5, 5.1), ncol =
  2), parallel = TRUE)
}
\arguments{
\item{spectra}{Data frame with spectra in columns and chemical shift in rows.
Colnames of this data frame correspond to pure metabolite names and rownames
to chemical shift grid (in p.p.m).}

\item{bin}{numeric specifying the bin width.}

\item{exclusion.areas}{definition domain of spectra to exclude (ppm).
By default, only the water region (4.5-5.1 ppm).}

\item{parallel}{If \code{TRUE}, apply function in parallel. Default to
\code{TRUE}.}
}
\value{
A data frame with normalised spectra in columns and buckets in rows
(bucket names correspond to the center of the bucket).
}
\description{
Apply a binning function on a data frame of spectra.
}
\examples{
current_path <- file.path(system.file("extdata", package = "ASICS"),
                          "spectra_example.txt")
spectra_data <- read.table(current_path, header = TRUE, row.names = 1)
spectra_bin <- binning(spectra_data, bin = 0.01)
}
