% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_pure_library}
\alias{create_pure_library}
\title{Create a pure library}
\usage{
create_pure_library(spectra, nb.protons, threshold = 1)
}
\arguments{
\item{spectra}{Data frame with spectra in column and chemical shift in row.
Colnames of this data frame correspond to pure metabolite names and rownames
to chemical shift grid (in p.p.m).}

\item{nb.protons}{Numeric vector of the number of protons of each pure
metabolite spectra contained in \code{spectra} data frame.}

\item{threshold}{Numeric value below which pure spectrum values are
considered null. Default to 1.}
}
\value{
A \linkS4class{PureLibrary} object with the newly created library.
}
\description{
Create a new pure library from a data frame containing spectra. Noise is
removed by thresholding each spectrum.
}
\examples{
pure_spectra <- import_spectra_bruker(system.file("extdata",
                                                  "example_library",
                                                   package = "ASICS"))
new_pure_library <- create_pure_library(pure_spectra,
                                        nb.protons = c(5, 4, 9, 9))

}
