% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{create_spectra}
\alias{create_spectra}
\title{Create a \linkS4class{Spectra} object}
\usage{
create_spectra(spectra)
}
\arguments{
\item{spectra}{Data frame with spectra in column and chemical shift in row.
Colnames of this data frame correspond to pure metabolite names and rownames
to chemical shift grid (in p.p.m).}
}
\value{
A \linkS4class{Spectra} object with spectra to quantify.
}
\description{
Create a new spectra object used for quantification.
}
\examples{
current_path <- system.file("extdata", "example_spectra", package = "ASICS")
spectra_data <- import_spectra_bruker(current_path)
spectra_obj <- create_spectra(spectra_data)
}
\seealso{
\linkS4class{Spectra}
}
