% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_concentrations.R
\name{extract_concentrations}
\alias{extract_concentrations}
\title{Extract concentrations}
\usage{
extract_concentrations(res_ASICS)
}
\arguments{
\item{res_ASICS}{result of ASICS_multiFiles functions}
}
\value{
A data frame containing relative concentrations of identified
metabolites for each spectrum
}
\description{
Description
}
\examples{
#Compute quantification on all spectra
res_multi <- ASICS_multiFiles(name.dir = system.file("extdata",
                                                     "example_spectra",
                                                     package = "ASICS"),
                           exclusion.areas = matrix(c(4.5,5.1,5.5,6.5),
                                                    ncol = 2, byrow = TRUE),
                           max.shift = 0.02, which.spectra = "last",
                           library.metabolites = NULL,
                           threshold.noise = 0.02, ncores = 1)

#Extract relative concentrations
quantification <- extract_concentrations(res_multi)
}
