% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractConcentrations.R
\name{extractConcentrations}
\alias{extractConcentrations}
\title{Extract concentrations}
\usage{
extractConcentrations(resASICS)
}
\arguments{
\item{resASICS}{result of ASICS or ASICS_multiFiles functions}
}
\description{
Description
}
\examples{
quantification <- extractConcentrations(result_all)
}
