% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_on_quantification.R
\name{format_for_analysis}
\alias{format_for_analysis}
\title{Format data for analysis}
\usage{
format_for_analysis(data, design = NULL, feature_info = NULL,
  zero.threshold = 1, outliers = NULL)
}
\arguments{
\item{data}{a data frame contaning omics data with samples in columns and
features of interest in rows (metabolites/genes...).}

\item{design}{a data frame describing the colums of \code{data} with at
least two columns and the first one corresponding to colnames of \code{data}.
Default to NULL (colnames of \code{data} is used for study design).}

\item{feature_info}{a data frame describing the rows of \code{data} with
at least two columns and the first one corresponding to rownames of
\code{data}. Default to NULL (rownames of \code{data} is used for feature
informations).}
}
\value{
An object of type \code{\link{SummarizedExperiment}} useful for
further analysis.
}
\description{
Create an object useful for analysis with omics data and informations
about the omics (metabolites/genes...) and study design.
}
