% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{importSpectraBruker}
\alias{importSpectraBruker}
\title{Import metabolomic spectra from Bruker files}
\usage{
importSpectraBruker(name.dir, which.spectra = "last",
  baseline.correction = TRUE, ppm.grid = NULL, sample.names = NULL,
  parallel = TRUE)
}
\arguments{
\item{name.dir}{Path of the folder containing one subfolder by sample. Each
subfolder contains the Bruker files of this sample.}

\item{which.spectra}{If there is more than one spectrum by sample, which is
the spectrum to import (either always the first one with \code{which.spectra
= "first"}, always the last one with \code{which.spectra = "last"} or a
vector containing the number of each spectra to import).
Default to \code{"last"}.}

\item{baseline.correction}{Logical. If \code{TRUE} a baseline correction is
applied for each spectrum (Wang et al (2013)). Default to \code{TRUE}.}

\item{ppm.grid}{Numeric vector of a unique grid (definition domain) for all
spectra (in ppm). Default to \code{NULL} (in which case, the default grid of
the pure library is used).}

\item{sample.names}{Character vector of sample names. Default to \code{NULL}
(in which case, folder names are used).}

\item{parallel}{Logical. If \code{TRUE}, the function is run in parallel.
Default to \code{TRUE}.}
}
\value{
A data frame with spectra in columns and chemical shifts (in ppm)
in rows.
}
\description{
Import spectra from Bruker files contained in a single folder. This folder
contains one subfolder for each sample. Spectra are baseline corrected
(optional) and normalised by the area under the curve during the importation.
}
\examples{
current_path <- system.file("extdata", "example_spectra", package = "ASICS")
spectra_data <- importSpectraBruker(current_path)
}
\references{
Wang, K.C., Wang, S.Y., Kuo, C.H., Tseng Y.J. (2013).
Distribution-based classification method for baseline correction of
metabolomic 1D proton nuclear magnetic resonance spectra.
\emph{Analytical Chemistry}, \strong{85}(2), 1231-1239.
}
\seealso{
\code{\link{baselineCorrection}} \code{\link{normalisation}}
\code{\link{alignment}}
}
