% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{normalisation}
\alias{normalisation}
\title{Normalisation}
\usage{
normalisation(spectra)
}
\arguments{
\item{spectra}{Data frame with spectra in columns and chemical shift in rows.
Colnames of this data frame correspond to pure metabolite names and rownames
to chemical shift grid (in p.p.m).}
}
\value{
A data frame with normalised spectra in columns and chemical shifts
(in p.p.m.) in rows.
}
\description{
Normalise a data frame of spectra by the area under the curve.
}
\examples{
current_path <- file.path(system.file("extdata", package = "ASICS"),
                          "spectra_example.txt")
spectra_data <- read.table(current_path, header = TRUE, row.names = 1)
spectra_norm <- normalisation(spectra_data)
}
