% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_on_quantification.R
\name{oplsda}
\alias{oplsda}
\title{Orthogonal projections to latent structures discriminant analysis (OPLS-DA)
on a \code{\link{SummarizedExperiment}} object}
\usage{
oplsda(analysis_data, condition, scale.unit = TRUE, cross.val = 1,
  thres.VIP = 1, orthoI = NA, type.data = "quantifications")
}
\arguments{
\item{analysis_data}{A \code{\link{SummarizedExperiment}} object obtained
with the \code{\link{formatForAnalysis}} function.}

\item{condition}{The name of the design variable (two level factor)
specifying the response to be explained.}

\item{scale.unit}{Logical. If \code{TRUE}, data are scaled to unit variance}

\item{cross.val}{Number of cross validation folds.}

\item{thres.VIP}{A number specifying the VIP threshold used to identify
influential variables.}

\item{orthoI}{Parameter of the \code{\link{opls}} function (package
\code{\link{ropls}}) specifying the number of orthogonal components. When set
to \code{NA}, OPLS is performed and the number of orthogonal components is
automatically tuned with cross-validation (with a maximum of 9 orthogonal
components). Default to \code{NA}.}

\item{type.data}{Type of data used for the analyses (\emph{e.g.,}
\code{"quantifications"}, \code{"buckets"}...). Default to
\code{"quantifications"}.}
}
\value{
A S4 object of class \linkS4class{AnalysisResults} containing OPLS-DA
results.
}
\description{
Perform an OPLS-DA with the function of the \code{\link{ropls}} package on a
\code{\link{SummarizedExperiment}} object obtained with the
\code{\link{formatForAnalysis}} function
}
\examples{
# Import quantification results
library(ASICSdata)
quantif_path <- system.file("extdata", "results_ASICS.txt",
                            package = "ASICSdata")
quantification <- read.table(quantif_path, header = TRUE, row.names = 1)

# Import design
design <- read.table(system.file("extdata", "design_diabete_example.txt",
                                 package = "ASICSdata"), header = TRUE)

# Create object for analysis and remove features with more than 25\% of zeros
analysis_obj <- formatForAnalysis(quantification,
                                  zero.threshold = 25, design = design)
res_oplsda <- oplsda(analysis_obj, "condition", orthoI = 1)

}
\references{
Thevenot, E.A., Roux, A., Xu, Y., Ezan, E., Junot, C. 2015.
Analysis of the human adult urinary metabolome variations with age, body mass
index and gender by implementing a comprehensive workflow for univariate and
OPLS statistical analyses. \emph{Journal of Proteome Research}.
\strong{14}:3322-3335.
}
\seealso{
\linkS4class{AnalysisResults}
}
