% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot ASICS results}
\usage{
plot_spectrum(res_ASICS, xmin = 0, xmax = 10, ymin = 0, ymax = NULL,
  add_metab = NULL)
}
\arguments{
\item{res_ASICS}{result of ASICS function}

\item{xmin, xmax}{x minimum and maximum limits}

\item{ymin, ymax}{y minimum and maximum limits}

\item{add_metab}{name of one metabolite to add to the plot}
}
\value{
Produces a plot of original and recontituted spectra.
}
\description{
Description
}
\examples{
#Compute quantification on one spectrum
result <- ASICS(path = system.file("extdata", "example_spectra",
                                   "AG_faq_Beck01", package = "ASICS"),
                exclusion.areas = matrix(c(4.5,5.1,5.5,6.5),
                                         ncol = 2, byrow = TRUE),
                max.shift = 0.02, which.spectra = "last",
                library.metabolites = NULL, threshold.noise = 0.02)

#Produces a plot of original, recontituted and Lactate spectra
plot_spectrum(result, xmin = 1, xmax = 1.5, ymax = 10, add_metab = "Lactate")
}
