% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectrum.R
\name{plotSpectrum}
\alias{plotSpectrum}
\title{Plot a spectrum}
\usage{
plotSpectrum(resASICS, xmin = 0, xmax = 10, ymin = 0, ymax = NULL,
  DecalageLib = 0, addMetab = NULL)
}
\arguments{
\item{resASICS}{result of ASICS function}

\item{xmin}{x minimum limit}

\item{xmax}{x maximum limit}

\item{ymin}{y minimum limit}

\item{ymax}{y maximum limit}

\item{DecalageLib}{chemical shift between library of pure spectra and complex mixture}

\item{addMetab}{name of one metabolite that we want to add the pure spectra}
}
\description{
Description
}
\examples{
plotSpectrum(result, xmin = 1, xmax = 1.5, ymax = 10, addMetab = "Lactate",
 DecalageLib = 0)
}
\keyword{NMR}
\keyword{metabolites}
\keyword{plot}
\keyword{quantification}
