% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pure_library}
\alias{pure_library}
\title{Pure spectra library}
\format{A list of 4 elements:
\describe{
  \item{name}{names of metabolites}
  \item{grid}{common grid for all spectra}
  \item{spectra}{a data frame with each pure metabolite spectrum in column}
  \item{nb_protons}{number of proton of each metabolite}
}}
\description{
A list containing 175 spectrums of pure metabolites.
}
