% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PureLibrary-class.R
\docType{class}
\name{PureLibrary-class}
\alias{PureLibrary-class}
\alias{get_nb_protons,PureLibrary-method}
\alias{[,PureLibrary,ANY,ANY,ANY-method}
\alias{[.PureLibrary}
\alias{c,PureLibrary-method}
\alias{c.PureLibrary}
\title{Class \code{PureLibrary}}
\usage{
\S4method{get_nb_protons}{PureLibrary}(object)

\S4method{[}{PureLibrary,ANY,ANY,ANY}(x, i)

\S4method{c}{PureLibrary}(x, ...)
}
\arguments{
\item{object}{an object of class \code{PureLibrary}}

\item{x}{an object of class \code{PureLibrary}}

\item{i}{indices specifying metabolites to extract}

\item{...}{currently not used}
}
\description{
Objects of class \code{PureLibrary} represent a set of pure metabolite NMR
spectra.
It an extension of class \linkS4class{Spectra} with an additional slot
(number of protons of each metabolite) needed for spectrum quantification.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_nb_protons}: extract number of protons of the \code{PureLibrary}
object.

\item \code{[}: extract some metabolites from a \code{PureLibrary}
object.

\item \code{c}: combine \code{PureLibrary} objects.
}}

\section{Slots}{

\describe{
\item{\code{nb.protons}}{Numeric vector of the number of protons of each pure
metabolite spectra.}
}}

\seealso{
\linkS4class{Spectra}
}
