% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R
\docType{class}
\name{Spectra-class}
\alias{Spectra-class}
\alias{get_sample_name,Spectra-method}
\alias{get_ppm_grid,Spectra-method}
\alias{get_spectra,Spectra-method}
\alias{show,Spectra-method}
\alias{show.Spectra}
\alias{summary,Spectra-method}
\alias{summary.Spectra}
\alias{[,Spectra,ANY,ANY,ANY-method}
\alias{[.Spectra}
\alias{length,Spectra-method}
\alias{length.Spectra}
\alias{c,Spectra-method}
\alias{c.Spectra}
\alias{plot,Spectra,ANY-method}
\alias{plot.Spectra}
\title{Class \code{Spectra}}
\usage{
\S4method{get_sample_name}{Spectra}(object)

\S4method{get_ppm_grid}{Spectra}(object)

\S4method{get_spectra}{Spectra}(object)

\S4method{show}{Spectra}(object)

\S4method{summary}{Spectra}(object)

\S4method{[}{Spectra,ANY,ANY,ANY}(x, i)

\S4method{length}{Spectra}(x)

\S4method{c}{Spectra}(x, ...)

\S4method{plot}{Spectra,ANY}(x, y, xlim = c(0.5, 10), ylim = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{Spectra}}

\item{x}{an object of class \code{Spectra}}

\item{i}{indices specifying elements to extract}

\item{...}{currently not used}

\item{y}{currently not used}

\item{xlim, ylim}{boundaries for x and y, respectively}
}
\description{
Objects of class \code{Spectra} represent a set of NMR spectra of one study.
It contains preprocessed spectra and can be created with the function
\code{\link{create_spectra}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_sample_name}: extract sample names of the \code{Spectra} object.

\item \code{get_ppm_grid}: extract p.p.m grid of the \code{Spectra} object.

\item \code{get_spectra}: extract spectra matrix of the \code{Spectra} object.

\item \code{show}: show a summary of the \code{Spectra} object.

\item \code{summary}: show a summary of the \code{Spectra} object.

\item \code{[}: extract some samples from a \code{Spectra} object.

\item \code{length}: number of samples in a \code{Spectra} object.

\item \code{c}: combine \code{Spectra} objects.

\item \code{plot}: plot all spectra (or a subset) on the same figure.
Legend only appears if there is less than ten spectra to plot.
}}

\section{Slots}{

\describe{
\item{\code{sample.name}}{Character vector of sample names.}

\item{\code{ppm.grid}}{Numeric vector of a unique grid (definition domain) for all
spectra (in ppm).}

\item{\code{spectra}}{Numeric matrix with all spectra in columns. Columns must be in
the same order as for \code{sample.name} and rows correspond to points of
\code{ppm.grid}.}
}}

