% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spectra-class.R, R/ASICSResults-class.R
\docType{methods}
\name{visualization-methods-spectra}
\alias{visualization-methods-spectra}
\alias{plot,Spectra,ANY-method}
\alias{plot.Spectra}
\alias{plot,ASICSResults,ANY-method}
\alias{plot.ASICSResults}
\title{Visualization methods}
\usage{
\S4method{plot}{Spectra,ANY}(x, y, xlim = c(0.5, 10), ylim = NULL, ...)

\S4method{plot}{ASICSResults,ANY}(x, y, idx = 1, xlim = c(0.5, 10),
  ylim = NULL, pure.library = NULL, add.metab = NULL, ...)
}
\arguments{
\item{x}{an object of class \linkS4class{Spectra},
\linkS4class{PureLibrary} or \linkS4class{ASICSResults}}

\item{y}{currently not used}

\item{xlim, ylim}{boundaries for x and y, respectively}

\item{...}{currently not used}

\item{idx}{index of the spectrum to plot. Default to 1.}

\item{pure.library}{pure library used for quantification. Default to
\code{NULL} (the library included in the package is used).}

\item{add.metab}{name of one metabolite to add to the plot. Default to
\code{NULL} (no pure spectrum added to the plot).}
}
\value{
\itemize{
\item A ggplot plot all spectra (or a subset) on the same figure for
\linkS4class{Spectra} and \linkS4class{PureLibrary} object.
\item A ggplot plot original and recomposed spectra of one sample in one
figure for \linkS4class{ASICSResults} object. In addition, one pure
metabolite spectrum (as provided in the reference library) and the deformed
one can be superimposed to the plot.
}
}
\description{
Methods available to plot one object in ASICS package.
}
\examples{
# Import data and create object
current_path <- system.file("extdata", "example_spectra", package = "ASICS")
spectra_data <- importSpectraBruker(current_path)
spectra_obj <- createSpectra(spectra_data)

# Plot the spectra
plot(spectra_obj)
}
