% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS.R
\name{ASICS}
\alias{ASICS}
\title{Automatic Statistical Identification in Complex Spectra}
\usage{
ASICS(path, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2, nrow = 1),
  max.shift = 0.02, which.spectra = "last", library.metabolites = NULL,
  threshold.noise = 0.02)
}
\arguments{
\item{path}{folder path of the Bruker files}

\item{exclusion.areas}{exclusion areas of the quantification}

\item{max.shift}{maximum chemical shift allowed (in ppm)}

\item{which.spectra}{if more than one spectra by sample, spectra to choose
(either "first", "last" or its number)}

\item{library.metabolites}{path of the library of standard if not the default
one}

\item{threshold.noise}{threshold for signal noise}
}
\value{
A object of type \link{resASICS-class}
}
\description{
Quantification of 1D 1H NMR spectrum with ASICS method using a library of 175
pure metabolite spectra. It is detailed in Tardivel et al. (2017).
}
\examples{
\dontrun{
result <- ASICS(path = system.file("extdata", "example_spectra",
                                   "AG_faq_Beck01", package = "ASICS"),
                exclusion.areas = matrix(c(4.5,5.1,5.5,6.5),
                                         ncol = 2, byrow = TRUE))
}
}
\references{
Tardivel, P., Canlet, C., Lefort, G., Tremblay-Franco, M.,
Debrauwer, L., Concordet, D., and Servien, R. (2017). ASICS: an automatic
method for identification and quantification of metabolites in complex 1D 1H
NMR spectra Metabolomics, 13: 109.
\url{https://doi.org/10.1007/s11306-017-1244-5}
}
\seealso{
\link{resASICS-class}
}
