% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS.R
\name{ASICS}
\alias{ASICS}
\title{Automatic Statistical Identification in Complex Spectra}
\usage{
ASICS(spectra_obj, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2),
  max.shift = 0.02, pure.library = NULL, threshold.noise = 0.02,
  seed = 1234, parallel = TRUE)
}
\arguments{
\item{spectra_obj}{an object of class \linkS4class{Spectra} obtained with the
function \link{preprocessing_spectra}}

\item{exclusion.areas}{definition domain of spectra to exclude (ppm)}

\item{max.shift}{maximum chemical shift allowed (in ppm)}

\item{pure.library}{an object of class \linkS4class{PureLibrary} containing
the references (pure metabolite spectra). If \code{NULL}, the library
included in the package is used}

\item{threshold.noise}{threshold for signal noise}

\item{seed}{random seed to control randomness in the algorithm (used in the
estimation of significativity of a given metabolite concentration)}

\item{parallel}{if \code{TRUE}, apply function in parallel. Default to
\code{TRUE}.}
}
\value{
An object of type \code{\link{ASICSResults}}.
}
\description{
Quantification of 1D 1H NMR spectra with ASICS method using a library of 175
pure metabolite spectra. The method is presented in Tardivel et al. (2017).
}
\references{
Tardivel P., Canlet C., Lefort G., Tremblay-Franco M., Debrauwer
L., Concordet D., Servien R. (2017). ASICS: an automatic method for
identification and quantification of metabolites in complex 1D 1H NMR
spectra. \emph{Metabolomics}, \strong{13}(10): 109.
\url{https://doi.org/10.1007/s11306-017-1244-5}
}
\seealso{
\code{\link{resASICS-class}} \code{\link{ASICS_multiFiles}}
\code{\link{pure_library}}
}
