% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS.R
\name{ASICS}
\alias{ASICS}
\title{Automatic Statistical Identification in Complex Spectra}
\usage{
ASICS(path, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2, nrow = 1),
  max.shift = 0.02, which.spectra = "last", library.metabolites = NULL,
  threshold.noise = 0.02)
}
\arguments{
\item{path}{folder path of the Bruker files}

\item{exclusion.areas}{exclusion areas of the quantification}

\item{max.shift}{maximum chemical shift allowed (in ppm)}

\item{which.spectra}{if more than one spectra by sample, spectra to choose
(either "first", "last" or its number)}

\item{library.metabolites}{path of the library of standard if not the default
one}

\item{threshold.noise}{threshold for signal noise}
}
\value{
A object of type resASICS
}
\description{
Description
}
\examples{
\dontrun{
result <- ASICS(path = system.file("extdata", "example_spectra",
                                   "AG_faq_Beck01", package = "ASICS"),
                exclusion.areas = matrix(c(4.5,5.1,5.5,6.5),
                                         ncol = 2, byrow = TRUE))
}
}
