% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS.R
\name{ASICS}
\alias{ASICS}
\title{Automatic Statistical Identification in Complex Spectra}
\usage{
ASICS(spectra_obj, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2),
  max.shift = 0.02, pure.library = NULL, threshold.noise = 0.02,
  seed = 1234, parallel = TRUE)
}
\arguments{
\item{spectra_obj}{An object of class \linkS4class{Spectra} obtained with the
function \link{createSpectra}.}

\item{exclusion.areas}{Definition domain of spectra that has to be excluded
for the quantification (ppm). By default, the water region is excluded
(4.5-5.1 ppm).}

\item{max.shift}{Maximum chemical shift allowed (in ppm). Default to 0.02.}

\item{pure.library}{An object of class \linkS4class{PureLibrary} containing
the reference spectra (pure metabolite spectra). If \code{NULL}, the library
included in the package (that contains 191 reference spectra) is used.}

\item{threshold.noise}{Threshold for signal noise. Default to 0.02.}

\item{seed}{Random seed to control randomness in the algorithm (used in the
estimation of the significativity of a given metabolite concentration).}

\item{parallel}{Logical. If \code{TRUE}, the function is run in parallel.
Default to \code{TRUE}.}
}
\value{
An object of type \linkS4class{ASICSResults} containing the
quantification results.
}
\description{
Quantification of 1D 1H NMR spectra with ASICS method using a library of
pure metabolite spectra. The method is presented in Tardivel et al. (2017).
}
\examples{
# Import data and create object
current_path <- system.file("extdata", "example_spectra", package = "ASICS")
spectra_data <- importSpectraBruker(current_path)
spectra_obj <- createSpectra(spectra_data)

# Estimation of relative quantification of Lactate and L-Alanine
to_exclude <- matrix(c(4.5, 10), ncol = 2)
pure_lib <- pure_library[pure_library@sample.name \%in\% c("Lactate",
"L-Alanine")]
resASICS <- ASICS(spectra_obj[1], exclusion.areas = to_exclude,
pure.library = pure_lib)
}
\references{
Tardivel P., Canlet C., Lefort G., Tremblay-Franco M., Debrauwer
L., Concordet D., Servien R. (2017). ASICS: an automatic method for
identification and quantification of metabolites in complex 1D 1H NMR
spectra. \emph{Metabolomics}, \strong{13}(10): 109.
\url{https://doi.org/10.1007/s11306-017-1244-5}
}
\seealso{
\linkS4class{ASICSResults} \code{\link{pure_library}}
\code{\link{createSpectra}}
}
