% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS.R
\name{ASICS}
\alias{ASICS}
\title{Automatic Statistical Identification in Complex Spectra}
\usage{
ASICS(path, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2, nrow = 1),
  max.shift = 0.02, which.spectra = "last", library.metabolites = NULL,
  threshold.noise = 0.02)
}
\arguments{
\item{path}{folder path of the Bruker files}

\item{exclusion.areas}{exclusion areas of the quantification}

\item{max.shift}{maximum chemical shift allowed (in ppm)}

\item{which.spectra}{if more than one spectra by sample, spectra to choose
(either "first", "last" or its number)}

\item{library.metabolites}{path of the library of standard if not the default
one}

\item{threshold.noise}{threshold for signal noise}
}
\value{
A list of 5 elements containing ASICS results:
\describe{
  \item{Present}{a data frame with identified metabolites and their relative
  concentrations}
  \item{Non_identified}{a data frame with non-identified metabolites and
  their identification thresholds}
  \item{Mixture}{original spectrum}
  \item{Estimated_mixture}{reconstituted spectrum with estimated
  concentrations}
  \item{Grid}{grid of the spectrum}
}
}
\description{
Description
}
\examples{
result <- ASICS(path = system.file("extdata", "example_spectra",
                                   "AG_faq_Beck01", package = "ASICS"),
                exclusion.areas = matrix(c(4.5,5.1,5.5,6.5),
                                         ncol = 2, byrow = TRUE),
                max.shift = 0.02, which.spectra = "last",
                library.metabolites = NULL, threshold.noise = 0.02)
}
