% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS_multiFiles.R
\name{ASICS_multiFiles}
\alias{ASICS_multiFiles}
\title{Automatic Statistical Identification in Complex Spectra for many files}
\usage{
ASICS_multiFiles(nameDir, ZoneAEnlever = matrix(c(4.5, 5.1), ncol = 2, nrow =
  1), DecalageMax = 0.021, DecalageLib = 0, ncores = 1, sample = "last",
  seed = 12345, libraryMetab = NULL, seuilBruit = 0.02)
}
\arguments{
\item{ZoneAEnlever}{exclusion areas to remove before the quantification}

\item{DecalageMax}{maximum chemical shift allowed}

\item{DecalageLib}{chemical shift between library of pure spectra and complex mixture}

\item{ncores}{number of cores to use}

\item{sample}{if more than one spectra by sample, spectra to choose (either "first",
"last" or its number}

\item{seed}{seed}

\item{libraryMetab}{path of the library of standard if not the default one}

\item{seuilBruit}{threshold noise}

\item{name}{folder path of the Bruker files}
}
\description{
Description
}
\examples{
ASICS_multiFiles(nameDir = "./spectres_exemple/",
ZoneAEnlever = matrix(c(4.5, 5.1), ncol = 2,
nrow = 1),
DecalageMax = 0.021, DecalageLib = -0.002,
ncores = 4, sample = "last", seed = 12345,
libraryMetab = NULL)
}
\keyword{NMR}
\keyword{metabolites}
\keyword{quantification}
