% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICS_multiFiles.R
\name{ASICS_multiFiles}
\alias{ASICS_multiFiles}
\title{Automatic Statistical Identification in Complex Spectra for many files}
\usage{
ASICS_multiFiles(name.dir, exclusion.areas = matrix(c(4.5, 5.1), ncol = 2,
  nrow = 1), max.shift = 0.02, which.spectra = "last",
  library.metabolites = NULL, threshold.noise = 0.02, ncores = 1)
}
\arguments{
\item{name.dir}{folder path of the Bruker files}

\item{exclusion.areas}{exclusion areas to remove before the quantification}

\item{max.shift}{maximum chemical shift allowed}

\item{which.spectra}{if more than one spectra by sample, spectra to choose
(either "first", "last" or its number)}

\item{library.metabolites}{path of the library of pure spectra if not the
default one}

\item{threshold.noise}{threshold for signal noise}

\item{ncores}{number of cores to use}
}
\value{
A list containing ASICS results for each spectrum.
}
\description{
Compute ASICS on multiple spectra of a folder.
}
\examples{
\dontrun{
res_multi <- ASICS_multiFiles(name.dir = system.file("extdata",
                                                     "example_spectra",
                                                     package = "ASICS"),
                           exclusion.areas = matrix(c(4.5,5.1,5.5,6.5),
                                                    ncol = 2, byrow = TRUE))
}
}
\seealso{
\link{ASICS}, \link{resASICS-class}
}
