% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASICSResults-class.R
\docType{class}
\name{ASICSResults-class}
\alias{ASICSResults-class}
\alias{get_recomposed_spectra,ASICSResults-method}
\alias{get_quantification,ASICSResults-method}
\alias{get_deformed_library,ASICSResults-method}
\alias{show,ASICSResults-method}
\alias{show.ASICSResults}
\alias{[,ASICSResults,ANY,ANY,ANY-method}
\alias{[.ASICSResults}
\alias{c,ASICSResults-method}
\alias{c.ASICSResults}
\title{Class \code{ASICSResults}}
\usage{
\S4method{get_recomposed_spectra}{ASICSResults}(object)

\S4method{get_quantification}{ASICSResults}(object)

\S4method{get_deformed_library}{ASICSResults}(object)

\S4method{show}{ASICSResults}(object)

\S4method{[}{ASICSResults,ANY,ANY,ANY}(x, i)

\S4method{c}{ASICSResults}(x, ...)
}
\arguments{
\item{object}{an object of class \code{ASICSResults}}

\item{i}{indices specifying samples to extract}
}
\description{
Objects of class \code{ASICSResults} represent results of ASICS for a set of
spectra.
It an extension of class \linkS4class{Spectra} with an additional slots for
quantification results.
}
\section{Methods (by generic)}{
\itemize{
\item \code{get_recomposed_spectra}: extract recomposed spectra matrix of the
\code{ASICSResults} object.

\item \code{get_quantification}: extract a data-frame with metabolite quantifications
computed with ASICS for each sample of the \code{ASICSResults} object.

\item \code{get_deformed_library}: extract a list of all deformed libraries of the
\code{ASICSResults} object.

\item \code{show}: show a summary of the \code{ASICSResults} object.

\item \code{[}: extract some samples from a \code{ASICSResults}
object.

\item \code{c}: combine \code{ASICSResults} objects.
}}

\section{Slots}{

\describe{
\item{\code{sample.name}}{Character vector of sample names.}

\item{\code{ppm.grid}}{Numeric vector of a unique grid (definition domain) for all
spectra (in ppm).}

\item{\code{spectra}}{Numeric matrix of original spectra in column.
Columns are in the same order as for \code{sample.name} and rows correspond
to points of \code{ppm.grid}.}

\item{\code{recomposed.spectra}}{Numeric matrix of recomposed spectra (in column)
with estimated concentrations. Columns are in the same order as for
\code{sample.name} and rows correspond to points of \code{ppm.grid}.}

\item{\code{quantification}}{Data-frame with identified metabolites and their
relative concentrations.}

\item{\code{deformed.library}}{a list of \linkS4class{PureLibrary} containing the
deformed libraries of each sample.}
}}

\seealso{
\linkS4class{Spectra}
}
