% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{baselineCorrection}
\alias{baselineCorrection}
\title{Baseline correction}
\usage{
baselineCorrection(spectra, ncores = 1, verbose = TRUE)
}
\arguments{
\item{spectra}{Data frame with spectra in columns and chemical shifts in
rows. Colnames of this data frame correspond to pure metabolite names and
rownames to chemical shift grid (in ppm).}

\item{ncores}{Number of cores used in parallel evaluation. Default to
\code{1}.}

\item{verbose}{A boolean value to allow print out process information.}
}
\value{
A data frame with baseline corrected spectra in columns and chemical
shifts (in ppm) in rows.
}
\description{
Apply a baseline correction to a spectra with the algorithm described in
Wang et al. (2013).
}
\examples{
current_path <- system.file("extdata", package = "ASICS")
spectra_data <- importSpectra(name.dir = current_path,
                     name.file = "spectra_example.txt", type.import = "txt")
spectra_base_cor <- baselineCorrection(spectra_data)
}
\references{
Wang, K.C., Wang, S.Y., Kuo, C.H., Tseng Y.J. (2013).
Distribution-based classification method for baseline correction of
metabolomic 1D proton nuclear magnetic resonance spectra.
\emph{Analytical Chemistry}, \strong{85}(2), 1231-1239.
}
