% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_on_quantification.R
\name{formatForAnalysis}
\alias{formatForAnalysis}
\title{Format data for analysis}
\usage{
formatForAnalysis(data, design = NULL, feature_info = NULL,
  zero.threshold = 100, zero.group = NULL, outliers = NULL)
}
\arguments{
\item{data}{a data frame contaning omics data with samples in columns and
features of interest in rows (metabolites/buckets...).}

\item{design}{a data frame describing the colums of \code{data} with at
least two columns and the first one corresponding to colnames of \code{data}.
Default to NULL (colnames of \code{data} is used for study design).}

\item{feature_info}{a data frame describing the rows of \code{data} with
at least two columns and the first one corresponding to rownames of
\code{data}. Default to NULL (rownames of \code{data} is used for feature
informations).}

\item{zero.threshold}{remove features with more than \code{zero.threshold}
percent of zeros. Default to \code{100}.}

\item{zero.group}{variable name of design data frame specifying the group
variable used to remove features with more than \code{zero.threshold}
percent of zeros. Default to \code{NULL}, no group.}

\item{outliers}{names of outliers to remove.}
}
\value{
An object of type \code{\link{SummarizedExperiment}} useful for
further analysis.
}
\description{
Create an useful object for analysis with omics data and informations
about the omics (metabolites/buckets...) and study design.
}
\examples{
# Import quantification results
library(ASICSdata)
quantif_path <- system.file("extdata", "results_ASICS.txt",
                            package = "ASICSdata")
quantification <- read.table(quantif_path, header = TRUE, row.names = 1)

# Import design
design <- read.table(system.file("extdata", "design_diabete_example.txt",
                                 package = "ASICSdata"), header = TRUE)

# Create object for analysis and remove features with more than 25\% of
#zeros
analysis_obj <- formatForAnalysis(quantification,
                                  design = design,
                                  zero.threshold = 25,
                                  zero.group = "condition")

}
