% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{importSpectraBruker}
\alias{importSpectraBruker}
\title{Import preprocessed metabolomic spectra from Bruker files}
\usage{
importSpectraBruker(name.dir, which.spectra = "first", ppm.grid = NULL,
  sample.names = NULL, ncores = 1, verbose = TRUE, ...)
}
\arguments{
\item{name.dir}{Path of the folder containing one subfolder by sample. Each
subfolder contains the Bruker files of this sample.}

\item{which.spectra}{If there is more than one spectrum by sample, which is
the spectrum to import (either always the first one with \code{which.spectra
= "first"}, always the last one with \code{which.spectra = "last"} or a
vector of length the number of spectra that specifies the number of each
spectrum to import).
Default to \code{"first"}.}

\item{ppm.grid}{Numeric vector of a unique grid (definition domain) for all
spectra (in ppm). Default to \code{NULL} (in which case, the default grid of
the pure library is used).}

\item{sample.names}{Character vector of sample names. Default to \code{NULL}
(in which case, folder names are used).}

\item{ncores}{Number of cores used in parallel evaluation. Default to
\code{1}.}

\item{verbose}{A boolean value to allow print out process information.}

\item{...}{Further arguments to be passed to the functions
\code{\link{alignSpectra}}, \code{\link{baselineCorrection}} or
\code{\link{normaliseSpectra}} for specifying the parameters of the
algorithm, if necessary.}
}
\value{
A data frame with spectra in columns and chemical shifts (in ppm)
in rows.
}
\description{
Import preprocessed spectra from Bruker files contained in a single folder.
This folder contains one subfolder for each sample.
(optional) Spectra are baseline corrected, aligned and normalised by the area
under the curve during the importation.
}
\details{
Some preprocessing steps are included during the importation. First, spectra
are baseline corrected if \code{baseline.correction = TRUE}. Then, all
spectrum definition domains are aligned to a unique one (either the one
specified in \code{ppm.grid} or the grid of the default library). Finally,
all spectra are normalised if \code{normalisation = TRUE}.
}
\examples{
current_path <- system.file("extdata", "example_spectra", package = "ASICS")
spectra_data <- importSpectraBruker(current_path)
}
\references{
Wang, K.C., Wang, S.Y., Kuo, C.H., Tseng Y.J. (2013).
Distribution-based classification method for baseline correction of
metabolomic 1D proton nuclear magnetic resonance spectra.
\emph{Analytical Chemistry}, \strong{85}(2), 1231-1239.
}
\seealso{
\code{\link{baselineCorrection}} \code{\link{normaliseSpectra}}
\code{\link{alignSpectra}}
}
