% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{importSpectraBruker}
\alias{importSpectraBruker}
\title{Import from Bruker files}
\usage{
importSpectraBruker(name.dir, which.spectra = "last",
  baseline.correction = TRUE, ppm.grid = NULL, sample.names = NULL,
  parallel = TRUE)
}
\arguments{
\item{name.dir}{Path of the folder containing one subfolder by sample. Each
subfolder contains Bruker files.}

\item{which.spectra}{If more than one spectra by sample, spectra to import
(either \code{"first"}, \code{"last"} or its number). Default to
\code{"last"}.}

\item{baseline.correction}{If \code{TRUE} apply a baseline correction for
each spectrum (Wang et al (2013)). Default to \code{TRUE}.}

\item{ppm.grid}{Numeric vector of a unique grid (definition domain) for all
spectra (in ppm). Default to \code{NULL} (grid of default pure library is
used).}

\item{sample.names}{Character vector of sample names. Default to \code{NULL}
(folder names are used).}

\item{parallel}{If \code{TRUE}, apply function in parallel. Default to
\code{TRUE}.}
}
\value{
A data frame with spectra in columns and chemical shifts (in p.p.m.)
in rows.
}
\description{
Import spectra from Bruker files contained in a single folder. This folder
contains subfolders for each sample. Spectra are baseline corrected
(optionnal) and normalized by the area under the curve.
}
\examples{
current_path <- system.file("extdata", "example_spectra", package = "ASICS")
spectra_data <- importSpectraBruker(current_path)
}
\references{
Wang, K.C., Wang, S.Y., Kuo, C.H., Tseng Y.J. (2013).
Distribution-based classification method for baseline correction of
metabolomic 1D proton nuclear magnetic resonance spectra.
\emph{Analytical Chemistry}, \strong{85}(2), 1231-1239.
}
\seealso{
\code{\link{baselineCorrection}} \code{\link{normalisation}}
\code{\link{alignment}}
}
