% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_on_quantification.R
\name{kruskalWallis}
\alias{kruskalWallis}
\title{Kruskal-Wallis rank sum tests on a \code{\link{SummarizedExperiment}} object}
\usage{
kruskalWallis(analysis_data, condition, correction = c("BH", "bonferroni",
  "BY", "fdr", "holm", "hochberg", "hommel"), alpha = 0.05,
  type.data = "quantifications")
}
\arguments{
\item{analysis_data}{a \code{\link{SummarizedExperiment}} object obtained 
with the \code{\link{formatForAnalysis}} function.}

\item{condition}{the name of the design variable (two level factor)
specifying the group of each sample.}

\item{correction}{p-value correction method, see \code{\link{p.adjust}}.
Default to \code{"BH"}.}

\item{alpha}{cutoff for adjusted p-values. Default to 0.05.}

\item{type.data}{type of data used for the analyses (\emph{e.g.,}
\code{"quantifications"}, \code{"buckets"}...). Default to
\code{"quantifications"}.}
}
\value{
A S4 object of class \linkS4class{AnalysisResults} containing test
results.
}
\description{
Perform Kruskal-Wallis tests on a \code{\link{SummarizedExperiment}} object
obtained with the \code{\link{formatForAnalysis}} function
}
\examples{
# Import quantification results
library(ASICSdata)
quantif_path <- system.file("extdata", "results_ASICS.txt",
                            package = "ASICSdata")
quantification <- read.table(quantif_path, header = TRUE, row.names = 1)

# Import design
design <- read.table(system.file("extdata", "design_diabete_example.txt",
                                 package = "ASICSdata"), header = TRUE)

# Create object for analysis and remove features with more than 25\% of zeros
analysis_obj <- formatForAnalysis(quantification,
                                    zero.threshold = 25, design = design)
res_tests <- kruskalWallis(analysis_obj, "condition")

}
\seealso{
\linkS4class{AnalysisResults}
}
