% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot original and recomposed spectra}
\usage{
plot_spectrum(ASICS.results, idx = 1, xlim = c(0, 10), ylim = NULL,
  pure.library = NULL, add.metab = NULL)
}
\arguments{
\item{ASICS.results}{an object of class \linkS4class{ASICSResults}.}

\item{idx}{index of the spectrum to plot. Default to 1.}

\item{xlim, ylim}{boundaries for x and y, respectively.}

\item{pure.library}{pure library used for quantification. Default to
\code{NULL} (the library included in the package is used).}

\item{add.metab}{name of one metabolite to add to the plot. Default to
\code{NULL} (no pure spectrum added to the plot).}
}
\value{
a ggplot plot original and recomposed spectra of one sample
in one figure. In addition, one pure metabolite spectrum (as
provided in the reference library) and the deformed one can be superimposed
to the plot.
}
\description{
Plot original spectrum and recomposed one obtained from ASICS quantification.
It is possible to add one pure metabolite spectrum.
}
