% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resASICS-class.R
\docType{class}
\name{resASICS-class}
\alias{resASICS-class}
\title{An S4 class to represent results of ASICS.}
\description{
An S4 class to represent results of ASICS.
}
\section{Slots}{

\describe{
\item{\code{original_mixture}}{original spectrum}

\item{\code{reconstituted_mixture}}{reconstituted spectrum with estimated
concentrations}

\item{\code{ppm_grid}}{grid of the spectrum (in p.p.m.)}

\item{\code{present_metabolites}}{a data frame with identified metabolites and their
relative concentrations}

\item{\code{non_identified_metabolites}}{a data frame with non-identified metabolites
and their identification thresholds}
}}

\note{
Slots can be accessed by accessor functions with the same name
(see \link{resASICS-getters})
}
