% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resASICS-class.R
\docType{methods}
\name{resASICS-methods}
\alias{resASICS-methods}
\alias{summary,resASICS-method}
\alias{summary.resASICS}
\alias{print,resASICS-method}
\alias{print.resASICS}
\alias{plot,resASICS-method}
\alias{plot.resASICS}
\title{S4 methods to represent results of ASICS.}
\usage{
\S4method{summary}{resASICS}(object, ...)

\S4method{print}{resASICS}(x)

\S4method{plot}{resASICS}(x, y, ..., xmin = 0, xmax = 10, ymin = 0,
  ymax = NULL, add_metab = NULL)
}
\arguments{
\item{object}{an object of class resASICS}

\item{...}{not used}

\item{x}{an object of class resASICS}

\item{y}{not used}

\item{xmin, xmax}{x minimum and maximum limits}

\item{ymin, ymax}{y minimum and maximum limits}

\item{add_metab}{name of one metabolite to add to the plot}
}
\description{
S4 methods to represent results of ASICS.
}
